+++ src/lang_utils/error_codes.ml
@@ -195,6 +195,11 @@ let error_codes : (string * string option) list =
     "M0219", None; (* Missing `transient` *)
     "M0220", None; (* Missing `persistent` *)
     "M0221", None; (* Failed to determine type for type pattern field *)
+    "M0225", None; (* A mixin cannot be used as an entry point *)
+    "M0226", None; (* Unknown mixin inclusion *)
+    "M0227", None; (* Non-actor include *)
+    "M0228", None; (* Only top-level mixins *)
+    "M0229", None; (* Non-var pattern for mixin import *)
   ]
 
 (** Message codes that can be both used as warnings and errors *)